commit on several files concerning the blank href references, image alt descriptions,unused declarations and unused imports
