--- submissions/jonasMwansa/react/react-app/src/DeliveryComponent.js
@@ -1,6 +1,0 @@
-import react from 'react'
-
-
-
-
-export default DeliveryComponent;
