+++ submissions/jonasMwansa/react/react-app/src/__tests__/CartPage.test.js
@@ -5,7 +5,7 @@ import CartPage from "../components/CartPage"
 const thisComponent = shallow(<CartPage />);
 
 it("renders without crashing", () => {
-    thisComponent 
+    thisComponent
 });
 
 it("test case: check header title", async () => {
