+++ submissions/jonasMwansa/react/react-app/src/components/CartContainer.js
@@ -1,41 +1,14 @@
-import { getByTitle } from '@testing-library/react';
-import React, { Component, useEffect, useState } from 'react';
-import image1 from '../image_assets/image1.svg'
-import image2 from '../image_assets/image3.svg'
-import image3 from '../image_assets/image4.svg'
+import React, { useEffect, useState } from 'react';
+
 import plus from '../image_assets/plus.svg'
 import minus from '../image_assets/Minus.svg'
 import cancel from '../image_assets/cancel.png'
-import Button from './Button';
 import scooter from '../image_assets/scooter 1.svg'
 
-const CartContainer = (props)=>{
+const CartContainer = ({products})=>{
 
     let total = 0;
     const fee= 1200;
-
-    const products = [
-        {
-            product_id:0,
-            product_name:'KFC - King Bucket',
-            image : image1,
-            price : 5000,
-            quantity: 1,
-        },
-        {
-            product_id:1,
-            product_name:'Refuel Max',
-            image: image2,
-            price: 1200,
-            quantity: 1,
-        },
-        {   product_id:2,
-            product_name:'Refuel Max',
-            image: image3,
-            price : 1200,
-            quantity: 1,
-        }
-    ]
 
 
 
@@ -44,7 +17,7 @@ const CartContainer = (props)=>{
     useEffect(()=> {
 
         setData(products)
-    },[])
+    },[products])
 
 
     //method to remove item from cart if quantity is only one item
@@ -114,7 +87,7 @@ const CartContainer = (props)=>{
                         <button id="remove-btn" onClick={()=> removeFromCart(item)}>
                             <img src={cancel} alt="cancel" id='cancel'></img>
                         </button>
-                        <img src={item.image} id="image" alt="product image"></img>
+                        <img src={item.image} id="image" alt="product item"></img>
 
                             <div className='title-buttons'>
                                 <p id='product-name'>{item.product_name}</p>
@@ -152,7 +125,7 @@ const CartContainer = (props)=>{
 
             <div className='delivery-details'>
                 <div className='scooter-delivery-title'>
-                    <img src={scooter}  id='scooter'></img>
+                    <img src={scooter}  id='scooter' alt='delivery man on scooter'></img>
                 </div>
                     <p id='delivery'>Delivery Fee</p>
                     <p id='delivery-price'>#{total===0? 0: fee}</p>
