--- submissions/jonasMwansa/react/react-app/src/components/CartItems.js
@@ -1,20 +1,0 @@
-import react from 'react';
-import image1 from '../image_assets/image1.svg'
-import image2 from '../image_assets/image3.svg'
-import image3 from '../image_assets/image4.svg'
-
-
-const CartItems= (props)=>{
-
-
-
-
-
-return(
-    <div>
-
-    </div>
-)
-}
-
-export default CartItems;
