+++ submissions/jonasMwansa/react/react-app/src/components/Navigation.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 
 
 const NavigationItem = (props)=>{
@@ -8,16 +7,16 @@ const NavigationItem = (props)=>{
         <div>
             <header>
                 <div id="logo">
-                    <a href="#"><h2>Food<span>Bag</span></h2></a>
+                    <a href="../../public/index.html"><h2>Food<span>Bag</span></h2></a>
                 </div>
             </header>
     {/*--header ends here--> */}
     
     <div class="navigation">
         <nav>
-            <a href="#" id="signup-link">Sign up</a>
-            <a href="#">Login</a>
-            <a href="#" id='cart-link'>Cart</a>
+            <a href="../../public/index.html" id="signup-link">Sign up</a>
+            <a href="../../public/index.html">Login</a>
+            <a href="../../public/index.html" id='cart-link'>Cart</a>
         </nav>
     </div>
     </div>
