+++ submissions/jonasMwansa/react/react-app/src/components/Sublink.js
@@ -1,4 +1,3 @@
-import React, { Component } from 'react';
 
 const Sublink = (props)=>{
     return(
