+++ submissions/jonasMwansa/react/react-app/src/index.css
@@ -96,15 +96,16 @@ nav a {
 }
 
 .content-display::after {
-  content: '';
+  content: "";
   position: absolute;
   width: 100%;
   height: 1px;
   left: 0;
   display: block;
   clear: both;
-  background-color: #C9C9C9;
+  background-color: #c9c9c9;
 }
+
 #product-name {
   font-size: 23px;
   font-weight: bold;
@@ -164,10 +165,11 @@ nav a {
   width: 200px;
   margin-left: 80px;
 }
+
 #delivery {
   font-size: 23px;
   font-weight: bold;
-  margin-left: 155px;
+  margin-left: 50px;
   margin-top: 100px;
 }
 
@@ -178,50 +180,48 @@ nav a {
   font-weight: bold;
 }
 
-.total-container{
-    display: flex;
-    justify-content: space-between;
+.total-container {
+  display: flex;
+  justify-content: space-between;
 }
 
 #total-title {
-    font-size: 30px;
+  font-size: 30px;
   font-weight: 600;
   margin-left: 100px;
   margin-top: 100px;
   color: #4e4d4d;
 }
 
-#total-price{
-    margin-right: 40px;
-    margin-top: 100px;
-    font-size: 23px;
-    font-weight: bold;
+#total-price {
+  margin-right: 40px;
+  margin-top: 100px;
+  font-size: 23px;
+  font-weight: bold;
 }
 
-.checkout-container{
-    padding-top: 30px;
-    margin: auto;
-    width: 50%;
+.checkout-container {
+  padding-top: 30px;
+  margin: auto;
+  width: 50%;
 }
 
-#checkout-button{
-    background: #AD4C4C;
-    border-radius: 5px;
-    width: 100%;
-    height: 60px;
-    font-weight: 600;
-    font-size: 16px;
-    color: white;
-    cursor: pointer;
+#checkout-button {
+  background: #ad4c4c;
+  border-radius: 5px;
+  width: 100%;
+  height: 60px;
+  font-weight: 600;
+  font-size: 16px;
+  color: white;
+  cursor: pointer;
 }
 
 #hidden {
-    display: none;
+  display: none;
 }
 
 @media (max-width: 600px) {
-  
-
   .navigation a {
     text-align: center;
     list-style: none;
@@ -230,34 +230,34 @@ nav a {
     padding: 5px;
     margin-right: 30px;
   }
-  
+
   .navigation a:nth-child(1) {
     border: 1px red solid;
     text-align: center;
     border-radius: 5px;
   }
-  
+
   nav a {
     text-align: center;
     text-decoration: none;
     color: gray;
   }
-  
+
   #signup-link,
   #cart-link {
     color: red;
   }
-  
+
   #heading {
     padding-top: 80px;
     padding-bottom: 30px;
     line-height: 0.5;
   }
-  
+
   .sub-title {
     padding-top: 50px;
   }
-  
+
   .sub-title a {
     color: gray;
     text-decoration: none;
@@ -267,7 +267,7 @@ nav a {
     display: flex;
     flex-direction: column;
   }
-  
+
   #image {
     height: 100px;
     width: 250px;
@@ -288,16 +288,16 @@ nav a {
     display: flex;
     justify-content: space-between;
   }
-  
+
   .content-display::after {
-    content: '';
+    content: "";
     position: absolute;
     width: 100%;
     height: 1px;
     left: 0;
     display: block;
     clear: both;
-    background-color: #C9C9C9;
+    background-color: #c9c9c9;
   }
 
   #product-name {
@@ -314,12 +314,12 @@ nav a {
     border: none;
     cursor: pointer;
   }
-  
+
   #plus-image {
     height: 40px;
     width: 40px;
   }
-  
+
   #minus-button {
     background-color: white;
     border: none;
@@ -333,7 +333,7 @@ nav a {
     padding-left: 4px;
     padding-right: 4px;
   }
-  
+
   #cancel {
     width: 15px;
     height: 15px;
@@ -345,7 +345,6 @@ nav a {
     height: 40px;
     cursor: pointer;
   }
-  
 
   #scooter {
     height: 150px;
@@ -359,31 +358,26 @@ nav a {
     margin-right: 50px;
     margin-top: 100px;
   }
-  
+
   #delivery-price {
     margin-right: 50px;
     margin-top: 100px;
     font-size: 16px;
     font-weight: bold;
   }
-  
 
-  
   #total-title {
-      font-size: 30px;
-      font-weight: 600;
-      margin-left: 80px;
-      margin-top: 100px;
-      color: #4e4d4d;
+    font-size: 30px;
+    font-weight: 600;
+    margin-left: 80px;
+    margin-top: 100px;
+    color: #4e4d4d;
   }
-  
-  #total-price{
-      margin-right: 40px;
-      margin-top: 100px;
-      font-size: 16px;
-      font-weight: bold;
+
+  #total-price {
+    margin-right: 40px;
+    margin-top: 100px;
+    font-size: 16px;
+    font-weight: bold;
   }
-  
-  
- 
 }
